+++ opencascade-hs/cpp/hs_BRepFilletAPI_MakeFillet.cpp
@@ -41,3 +41,7 @@ int hs_BRepFilletAPI_MakeFillet_nbEdges(BRepFilletAPI_MakeFillet * builder, int 
 TopoDS_Edge * hs_BRepFilletAPI_MakeFillet_edge(BRepFilletAPI_MakeFillet * builder, int contourIndex, int edgeIndex){
     return new TopoDS_Edge(builder->Edge(contourIndex, edgeIndex));
 }
+
+void hs_BRepFilletAPI_MakeFillet_remove(BRepFilletAPI_MakeFillet * builder, TopoDS_Edge * edge){
+    builder->Remove(*edge);
+}