+++ opencascade-hs/cpp/hs_BRepFilletAPI_MakeFillet.h
@@ -27,6 +27,8 @@ int hs_BRepFilletAPI_MakeFillet_nbEdges(BRepFilletAPI_MakeFillet * builder, int 
 
 TopoDS_Edge * hs_BRepFilletAPI_MakeFillet_edge(BRepFilletAPI_MakeFillet * builder, int contourIndex, int edgeIndex);
 
+void hs_BRepFilletAPI_MakeFillet_remove(BRepFilletAPI_MakeFillet * builder, TopoDS_Edge * edge);
+
 #ifdef __cplusplus
 }
 #endif